pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "splice-xai"
version = "0.1.0"
authors = [{ name = "Gunel Aghakishiyeva", email = "gunel.aghakishiyeva@duke.edu" }]
description = "SPLICE: Semantically Plausible Localized Inpainting for Context-preserving Explanations in Ecology"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"

dependencies = [
  "numpy>=1.26",
  "Pillow>=9.0.0",
  "matplotlib>=3.7",
  "requests>=2.32.0",
  "python-dotenv>=1.0.0",
  # heavy stacks are guarded in requirements.txt; okay to leave them out here
  "ultralytics>=8.3.0",
  "replicate>=1.0.0",
]

classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/duke-trust-lab/splice-xai"
Repository = "https://github.com/duke-trust-lab/splice-xai"
Issues = "https://github.com/duke-trust-lab/splice-xai/issues"
Documentation = "https://github.com/duke-trust-lab/splice-xai/README.md"

[project.scripts]
splice-xai = "splice_xai.cli.main:main"

[tool.setuptools.packages.find]
where = ["src"]

